Feat: Fix sorting of series where new seasons only append the year
AniDB will sometime name split-cour or individual seasons of a long running series as "Series (year)". 

For example:
- Tonikaku Kawaii
- Tonikaku Kawaii (2023)

https://anidb.net/anime/15421
https://anidb.net/anime/16961

To aid in searching, we append the series' alt-title to the end of the Sort Title in brackets. However, when sorting, the parenthesis will sort before the the square bracket does. This results in weird sorting for titles like the above. 

Plex will sort: Tonikaku Kawaii (2023) -> Tonikaku Kawaii

This simply appends an exclamation mark prior to appending the alt-title. This way, shows without the year appended will always sort first thanks to the exclamation mark having higher precedence.